requirements.txt
pytest==8.3.5
pytest-playwright==0.6.2
pytest-asyncio==0.24.0
pytest-html==4.1.1
pytest-xdist==3.6.1
playwright==1.49.0
python-dotenv==1.0.1
allure-pytest==2.13.5
faker==33.1.0
